created curses screen for sim
